+++ CHANGELOG.md
@@ -10,11 +10,12 @@
 * Can handle missing/no-data class labels (#45)
 * Adds CLI command `fill` for filling missing data (#49)
 * Adds CLI command `predict` for predicting complete layers (#50)
-* Adds `codes` plots showing a 2D array of small line, pie or rose charts, one for each node (#52, #53, #53)
+* Adds `codes` plots showing a 2D array of small line, pie, rose or image plot, one for each node (#52, #53, #53, #54, #55)
 
 ### Documentation
 
 * Examples are included in release downloads (#51)
+* Adds hand-written digits example (#55)
 
 ### Bugfixes